Split CurriculumVitaeTest file
